requirements.txt
streamlit==1.51.0
fastapi==0.122.0
pydantic==2.12.5
uvicorn==0.38.0
mysql-connector-python==9.5.0
pandas==2.3.3
requests==2.32.5
pytest==9.0.1
